// src/app/api/swagger/models/api-application-available-operation-dto.ts
/* tslint:disable */
/* eslint-disable */
export interface ApiApplicationAvailableOperationDto {
  IsApplicationClarificationResponseAllowed?: boolean;
  IsApproveAllowed?: boolean;
  IsAttachPaperDealDocumentsAllowed?: boolean;
  IsChangeContractFormAllowed?: boolean;
  IsChangeDealAllowed?: boolean;
  IsChangePayMethodAllowed?: boolean;
  IsChangeToPaperDealFormResponseAllowed?: boolean;
  IsCreateAdditionalAgreementsAllowed?: boolean;
  IsDealAgreementAllowed?: boolean;
  IsDealAgreementApproveAllowed?: boolean;
  IsDealExecutionOrTerminationAllowed?: boolean;
  IsDealPaperFormAllowed?: boolean;
  IsDifferencesProtocolAllowed?: boolean;
  IsRejectAdditionalAgreementAllowed?: boolean;
  IsRevokeApplicationAllowed?: boolean;
  IsRevokeDealAllowed?: boolean;
  IsSendContractAllowed?: boolean;
  IsSendDealToEisAllowed?: boolean;
  IsSignAdditionalAgreementAllowed?: boolean;
  IsSupplierDataClarificationAllowed?: boolean;
}