// src/app/api/ui/models/application-available-operation.interface.ts
export interface ApplicationAvailableOperation {
  isApplicationClarificationResponseAllowed: boolean;
  isApproveAllowed: boolean;
  isAttachPaperDealDocumentsAllowed: boolean;
  isChangeContractFormAllowed: boolean;
  isChangeDealAllowed: boolean;
  isChangePayMethodAllowed: boolean;
  isChangeToPaperDealFormResponseAllowed: boolean;
  isCreateAdditionalAgreementsAllowed: boolean;
  isDealAgreementAllowed: boolean;
  isDealAgreementApproveAllowed: boolean;
  isDealExecutionOrTerminationAllowed: boolean;
  isDealPaperFormAllowed: boolean;
  isDifferencesProtocolAllowed: boolean;
  isRejectAdditionalAgreementAllowed: boolean;
  isRevokeApplicationAllowed: boolean;
  isRevokeDealAllowed: boolean;
  isSendContractAllowed: boolean;
  isSendDealToEisAllowed: boolean;
  isSignAdditionalAgreementAllowed: boolean;
  isSupplierDataClarificationAllowed: boolean;
}